Fix - don't call API endpoint if request is not valid. (#759)
Remap function is testing for several cases of invalid requests. In some cases it "return"s when validation fails. In other cases it just adds an error response, and continue to execute. This cause endpoints to be called for failed requests (unauthorized, unauthenticated, invalid params etc').

I added a validation flag, fliping it in relevant places, and testing it at the end, before calling the endpoint.